Fixes #130 (Instead will crash with 17 generated Surveys!)
There was a tag-key mixup between IslandInfo and SurveyScan, causing
crashes with pre-17 generated Surveys.

This also makes surveys from 17 incompatible and will result in them
crashing the game instead, but this is a nessesary evil.

Note for myself: Make users aware of this in the log, and tell them to
get rid of any surveys generated with 17!!!