Coker and HPRU restoration
The HPRU already was in a properly working state, just the Coker has some minor issues.
Those are fixed now though.

More specifically a hopper pushing items into it caused all slots to be filled with bitumen, which is of course not how it should work.

CokerUnitLogic class was moved to its own package and the CokingChamber has been put into its own file as well.

The rest is mostly cleanup.